backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2

# Database
motor==3.3.1
pymongo==4.5.0
certifi==2024.12.14

# Authentication & Security
bcrypt==4.1.3
PyJWT==2.10.1
slowapi==0.1.9

# Environment & Configuration
python-dotenv==1.1.1

# HTTP Client
requests==2.32.5
httpx==0.26.0

# Email Validation
email-validator==2.3.0

# Data Validation
pydantic==2.12.0
pydantic_core==2.41.1

# CORS
# (included in starlette)

# Async Support
anyio==4.11.0

# Utils
python-multipart==0.0.20

# AI/ML Dependencies
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
google-generativeai==0.8.3

# Background Jobs
APScheduler==3.10.4

# Monitoring & Logging (Phase 4)
psutil==5.9.8

# Payment (PayOS)
# No SDK required, using direct API calls via requests

# Email Service (Resend)
resend==2.21.0
